Makes sure PYTHONPATH is set up and unzip isnt repeated

+++ stage.sh
@@ -68,26 +68,28 @@ stage util/ddlog.jar                                              lib/
 stage ddlib/ddlib                                                 lib/python/
 
 # DimmWitted sampler
+if ! [[ -e "$STAGE_DIR"/util/sampler-dw ]]; then
 case $(uname) in
 Linux)
     stage util/format_converter_linux                             util/format_converter
+    unzip -o util/sampler-dw-linux-lib.zip                     -d "$STAGE_DIR"/lib/
     #stage sampler/dw                                             util/sampler-dw # TODO
     stage util/sampler-dw-linux                                   util/
     stage util/sampler-dw-linux.sh                                util/
     ln -sfn sampler-dw-linux.sh                                   "$STAGE_DIR"/util/sampler-dw
-    unzip -o util/sampler-dw-linux-lib.zip                     -d "$STAGE_DIR"/lib/
     ;;
 Darwin)
     stage util/format_converter_mac                               util/format_converter
+    ditto -xk util/sampler-dw-mac-lib.zip                         "$STAGE_DIR"/lib/
     stage util/sampler-dw-mac                                     util/
     stage util/sampler-dw-mac.sh                                  util/
     ln -sfn sampler-dw-mac.sh                                     "$STAGE_DIR"/util/sampler-dw
-    ditto -xk util/sampler-dw-mac-lib.zip                         "$STAGE_DIR"/lib/
     ;;
 *)
     echo >&2 "$(uname): Unsupported OS"; false
     ;;
 esac
+fi
 
 # piggy extractor helper
 stage util/piggy_prepare.py                                       util/
